Rename helper to domain-logic
While the name 'domain-logic' is arguable, the intention is to
represent that scoring is the core logic of the business domain.
The name 'helper' is more of set of utility functions that is more
generic and help reducing repetitive tasks.
